+++ src/apis/mocks/cartItemsData/getCartItemsHandler.ts
@@ -1,0 +1,6 @@
+import { HttpResponse } from "msw";
+import { cartItemsMockData } from "./cartItemsMockData";
+
+export const getCartItemsHandler = () => {
+  return HttpResponse.json(cartItemsMockData, { status: 200 });
+};
